moving things around for better separation of concerns
