+++ stack/client/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.24",
   "private": true,
   "dependencies": {
+    "@mysticrobots/common": "file:../common",
     "@reduxjs/toolkit": "^1.3.5",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
