+++ stack/common/models/index.js
@@ -1,0 +1,9 @@
+const Challenge = require('./Challenge')
+const ChallengePuzzle = require('./ChallengePuzzle')
+const Score = require('./Score')
+
+module.exports = {
+    Challenge,
+    ChallengePuzzle,
+    Score,
+}
