+++ stack/common/package.json
@@ -1,10 +1,12 @@
 {
-  "name": "common",
+  "name": "@mysticrobots/common",
   "version": "0.1.3",
   "private": true,
   "main": "index.js",
   "directories": {
     "lib": "lib",
+    "models": "models",
+    "services": "services",
     "test": "test"
   },
   "files": [
