+++ stack/server/routes/challenges/get_challenge_by_id.js
@@ -1,14 +1,14 @@
-const challenges = require('../services/challenges')
+const challenges = require('../../../common/services/challenges')
 
 const get_challenge_by_id = async (req, res) => {
-    const id = req.params.id
+    const challengeId = req.params.challengeId
 
-    if(!req.params.id) {
+    if(!req.params.challengeId) {
         res.status(404).send("No challenge found")
         return
     }
  
-    const data = await challenges.get_challenge_by_id({ challengeId: req.params.id })
+    const data = await challenges.get_challenge_by_id({ challengeId: req.params.challengeId })
 
     if(!data) {
         res.status(404).send("No challenge found")
