+++ stack/server/routes/challenges/get_challenges.js
@@ -1,4 +1,4 @@
-const challenges = require('../services/challenges')
+const challenges = require('../../../common/services/challenges')
 
 const get_challenges = async (req, res) => {
     const data = await challenges.get_challenges({ latest: req.query.latest })
