+++ stack/server/routes/puzzles/get_puzzle_by_id.js
