+++ stack/server/routes/scores/get_scores_by_challenge_id.js
